+++ lib/src/mixin/dispose_bag.mixin.dart
@@ -1,0 +1,13 @@
+import 'dart:async';
+
+import 'package:flutter/cupertino.dart';
+
+mixin DisposeBag<T extends StatefulWidget> on State<T> {
+  List<StreamSubscription> disposeBag = [];
+
+  @override
+  void dispose() {
+    disposeBag.forEach((it) => it.cancel());
+    super.dispose();
+  }
+}